requirements.txt
python-binance
python-dotenv
rich
questionary
streamlit
portalocker
flask